revision/pyproject.toml
[project]
name = "revision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "logtail-python>=0.3.4",
    "pika>=1.3.2",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
]
